@@ -28,7 +29,15 @@ repositories {
 	maven { url 'http://repo.spring.io/libs-milestone' }
 }
 
-sourceCompatibility = targetCompatibility = 1.7
+compileJava {
+	sourceCompatibility = 1.7
+	targetCompatibility = 1.7
+}
+
+compileTestJava {
+	sourceCompatibility = 1.8
+	targetCompatibility = 1.8
+}
 
 if (project.hasProperty('platformVersion')) {
 	apply plugin: 'spring-io'
