@@ -1,33 +1,29 @@
 /*
  * Copyright 2016 the original author or authors.
  *
- *  Licensed under the Apache License, Version 2.0 (the "License");
- *  you may not use this file except in compliance with the License.
- *  You may obtain a copy of the License at
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
  *
- *       http://www.apache.org/licenses/LICENSE-2.0
+ *      http://www.apache.org/licenses/LICENSE-2.0
  *
- *  Unless required by applicable law or agreed to in writing, software
- *  distributed under the License is distributed on an "AS IS" BASIS,
- *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  See the License for the specific language governing permissions and
- *  limitations under the License.
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
  */
 
 package org.springframework.integration.aws.outbound;
 
-import static org.hamcrest.Matchers.instanceOf;
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertNotNull;
-import static org.junit.Assert.assertNull;
-import static org.junit.Assert.assertTrue;
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.mockito.BDDMockito.given;
+import static org.mockito.BDDMockito.willAnswer;
 import static org.mockito.Matchers.any;
 import static org.mockito.Mockito.atLeastOnce;
-import static org.mockito.Mockito.doAnswer;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.spy;
 import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.when;
 
 import java.io.ByteArrayInputStream;
 import java.io.File;
@@ -36,7 +32,6 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.util.Arrays;
 import java.util.Collections;
-import java.util.Comparator;
 import java.util.HashMap;
 import java.util.LinkedList;
 import java.util.List;
@@ -45,14 +40,11 @@ import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.TimeUnit;
 
 import org.apache.http.client.methods.HttpRequestBase;
-import org.junit.Assert;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.TemporaryFolder;
 import org.junit.runner.RunWith;
 import org.mockito.ArgumentCaptor;
-import org.mockito.invocation.InvocationOnMock;
-import org.mockito.stubbing.Answer;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
@@ -149,21 +141,21 @@ public class S3MessageHandlerTests {
 		verify(this.amazonS3, atLeastOnce()).putObject(putObjectRequestArgumentCaptor.capture());
 
 		PutObjectRequest putObjectRequest = putObjectRequestArgumentCaptor.getValue();
-		assertEquals("myBucket", putObjectRequest.getBucketName());
-		assertEquals("foo.mp3", putObjectRequest.getKey());
-		assertNull(putObjectRequest.getFile());
-		assertNotNull(putObjectRequest.getInputStream());
+		assertThat(putObjectRequest.getBucketName()).isEqualTo("myBucket");
+		assertThat(putObjectRequest.getKey()).isEqualTo("foo.mp3");
+		assertThat(putObjectRequest.getFile()).isNull();
+		assertThat(putObjectRequest.getInputStream()).isNotNull();
 
 		ObjectMetadata metadata = putObjectRequest.getMetadata();
-		assertEquals(Md5Utils.md5AsBase64(file), metadata.getContentMD5());
-		assertEquals(0, metadata.getContentLength());
-		assertEquals("audio/mpeg", metadata.getContentType());
+		assertThat(metadata.getContentMD5()).isEqualTo(Md5Utils.md5AsBase64(file));
+		assertThat(metadata.getContentLength()).isEqualTo(0);
+		assertThat(metadata.getContentType()).isEqualTo("audio/mpeg");
 
 		ProgressListener listener = putObjectRequest.getGeneralProgressListener();
 		S3ProgressPublisher.publishProgress(listener, ProgressEventType.TRANSFER_COMPLETED_EVENT);
 
-		assertTrue(this.transferCompletedLatch.await(10, TimeUnit.SECONDS));
-		assertTrue(this.aclLatch.await(10, TimeUnit.SECONDS));
+		assertThat(this.transferCompletedLatch.await(10, TimeUnit.SECONDS)).isTrue();
+		assertThat(this.aclLatch.await(10, TimeUnit.SECONDS)).isTrue();
 
 		ArgumentCaptor<SetObjectAclRequest> setObjectAclRequestArgumentCaptor =
 				ArgumentCaptor.forClass(SetObjectAclRequest.class);
@@ -171,10 +163,10 @@ public class S3MessageHandlerTests {
 
 		SetObjectAclRequest setObjectAclRequest = setObjectAclRequestArgumentCaptor.getValue();
 
-		assertEquals("myBucket", setObjectAclRequest.getBucketName());
-		assertEquals("foo.mp3", setObjectAclRequest.getKey());
-		assertNull(setObjectAclRequest.getAcl());
-		assertEquals(CannedAccessControlList.PublicReadWrite, setObjectAclRequest.getCannedAcl());
+		assertThat(setObjectAclRequest.getBucketName()).isEqualTo("myBucket");
+		assertThat(setObjectAclRequest.getKey()).isEqualTo("foo.mp3");
+		assertThat(setObjectAclRequest.getAcl()).isNull();
+		assertThat(setObjectAclRequest.getCannedAcl()).isEqualTo(CannedAccessControlList.PublicReadWrite);
 	}
 
 	@Test
@@ -192,16 +184,16 @@ public class S3MessageHandlerTests {
 		verify(this.amazonS3, atLeastOnce()).putObject(putObjectRequestArgumentCaptor.capture());
 
 		PutObjectRequest putObjectRequest = putObjectRequestArgumentCaptor.getValue();
-		assertEquals("myBucket", putObjectRequest.getBucketName());
-		assertEquals("myStream", putObjectRequest.getKey());
-		assertNull(putObjectRequest.getFile());
-		assertNotNull(putObjectRequest.getInputStream());
+		assertThat(putObjectRequest.getBucketName()).isEqualTo("myBucket");
+		assertThat(putObjectRequest.getKey()).isEqualTo("myStream");
+		assertThat(putObjectRequest.getFile()).isNull();
+		assertThat(putObjectRequest.getInputStream()).isNotNull();
 
 		ObjectMetadata metadata = putObjectRequest.getMetadata();
-		assertEquals(Md5Utils.md5AsBase64(payload), metadata.getContentMD5());
-		assertEquals(1, metadata.getContentLength());
-		assertEquals(MediaType.APPLICATION_JSON_VALUE, metadata.getContentType());
-		assertEquals("test.json", metadata.getContentDisposition());
+		assertThat(metadata.getContentMD5()).isEqualTo(Md5Utils.md5AsBase64(payload));
+		assertThat(metadata.getContentLength()).isEqualTo(1);
+		assertThat(metadata.getContentType()).isEqualTo(MediaType.APPLICATION_JSON_VALUE);
+		assertThat(metadata.getContentDisposition()).isEqualTo("test.json");
 	}
 
 	@Test
@@ -214,26 +206,19 @@ public class S3MessageHandlerTests {
 		this.s3SendChannel.send(message);
 
 		File[] fileArray = directoryForDownload.listFiles();
-		assertNotNull(fileArray);
-		assertEquals(2, fileArray.length);
+		assertThat(fileArray).isNotNull();
+		assertThat(fileArray.length).isEqualTo(2);
 
 		List<File> files = Arrays.asList(fileArray);
-		Collections.sort(files, new Comparator<File>() {
-
-			@Override
-			public int compare(File o1, File o2) {
-				return o1.getName().compareTo(o2.getName());
-			}
-
-		});
+		Collections.sort(files, (o1, o2) -> o1.getName().compareTo(o2.getName()));
 
 		File file1 = files.get(0);
-		assertEquals("bar", file1.getName());
-		assertEquals("bb", FileCopyUtils.copyToString(new FileReader(file1)));
+		assertThat(file1.getName()).isEqualTo("bar");
+		assertThat(FileCopyUtils.copyToString(new FileReader(file1))).isEqualTo("bb");
 
 		File file2 = files.get(1);
-		assertEquals("foo", file2.getName());
-		assertEquals("f", FileCopyUtils.copyToString(new FileReader(file2)));
+		assertThat(file2.getName()).isEqualTo("foo");
+		assertThat(FileCopyUtils.copyToString(new FileReader(file2))).isEqualTo("f");
 	}
 
 	@Test
@@ -245,15 +230,15 @@ public class S3MessageHandlerTests {
 		this.s3ProcessChannel.send(new GenericMessage<>(payload));
 
 		Message<?> receive = this.s3ReplyChannel.receive(10000);
-		assertNotNull(receive);
+		assertThat(receive).isNotNull();
 
-		Assert.assertThat(receive.getPayload(), instanceOf(Copy.class));
+		assertThat(receive.getPayload()).isInstanceOf(Copy.class);
 		Copy copy = (Copy) receive.getPayload();
-		assertEquals("Copying object from myBucket/mySource to theirBucket/theirTarget", copy.getDescription());
+		assertThat(copy.getDescription()).isEqualTo("Copying object from myBucket/mySource to theirBucket/theirTarget");
 
 		copy.waitForCompletion();
 
-		assertEquals(Transfer.TransferState.Completed, copy.getState());
+		assertThat(copy.getState()).isEqualTo(Transfer.TransferState.Completed);
 	}
 
 	@Configuration
@@ -264,9 +249,9 @@ public class S3MessageHandlerTests {
 		public AmazonS3 amazonS3() {
 			AmazonS3 amazonS3 = mock(AmazonS3.class);
 
-			when(amazonS3.putObject(any(PutObjectRequest.class))).thenReturn(new PutObjectResult());
-			when(amazonS3.getObjectMetadata(any(GetObjectMetadataRequest.class))).thenReturn(new ObjectMetadata());
-			when(amazonS3.copyObject(any(CopyObjectRequest.class))).thenReturn(new CopyObjectResult());
+			given(amazonS3.putObject(any(PutObjectRequest.class))).willReturn(new PutObjectResult());
+			given(amazonS3.getObjectMetadata(any(GetObjectMetadataRequest.class))).willReturn(new ObjectMetadata());
+			given(amazonS3.copyObject(any(CopyObjectRequest.class))).willReturn(new CopyObjectResult());
 
 			ObjectListing objectListing = spy(new ObjectListing());
 
@@ -284,8 +269,8 @@ public class S3MessageHandlerTests {
 			fileSummary2.setSize(2);
 			s3ObjectSummaries.add(fileSummary2);
 
-			when(objectListing.getObjectSummaries()).thenReturn(s3ObjectSummaries);
-			when(amazonS3.listObjects(any(ListObjectsRequest.class))).thenReturn(objectListing);
+			given(objectListing.getObjectSummaries()).willReturn(s3ObjectSummaries);
+			given(amazonS3.listObjects(any(ListObjectsRequest.class))).willReturn(objectListing);
 
 			final S3Object file1 = new S3Object();
 			file1.setBucketName("myBucket");
@@ -317,34 +302,28 @@ public class S3MessageHandlerTests {
 				// no-op
 			}
 
-			doAnswer(new Answer<S3Object>() {
-
-				@Override
-				public S3Object answer(InvocationOnMock invocation) throws Throwable {
-					GetObjectRequest getObjectRequest = (GetObjectRequest) invocation.getArguments()[0];
-					String key = getObjectRequest.getKey();
-					if ("foo".equals(key)) {
-						return file1;
-					}
-					else if ("bar".equals(key)) {
-						return file2;
-					}
-					else {
-						return (S3Object) invocation.callRealMethod();
-					}
+			willAnswer(invocation -> {
+				GetObjectRequest getObjectRequest = (GetObjectRequest) invocation.getArguments()[0];
+				String key = getObjectRequest.getKey();
+				if ("foo".equals(key)) {
+					return file1;
 				}
-
-			}).when(amazonS3).getObject(any(GetObjectRequest.class));
-
-			doAnswer(new Answer<Object>() {
-
-				@Override
-				public Object answer(InvocationOnMock invocation) throws Throwable {
-					aclLatch().countDown();
-					return null;
+				else if ("bar".equals(key)) {
+					return file2;
 				}
+				else {
+					return invocation.callRealMethod();
+				}
+			})
+					.given(amazonS3)
+					.getObject(any(GetObjectRequest.class));
 
-			}).when(amazonS3).setObjectAcl(any(SetObjectAclRequest.class));
+			willAnswer(invocation -> {
+				aclLatch().countDown();
+				return null;
+			})
+					.given(amazonS3)
+					.setObjectAcl(any(SetObjectAclRequest.class));
 
 			return amazonS3;
 		}
@@ -388,17 +367,12 @@ public class S3MessageHandlerTests {
 					PARSER.parseExpression("payload instanceof T(java.io.File) ? payload.name : headers.key");
 			s3MessageHandler.setKeyExpression(keyExpression);
 			s3MessageHandler.setObjectAclExpression(new ValueExpression<>(CannedAccessControlList.PublicReadWrite));
-			s3MessageHandler.setUploadMetadataProvider(new S3MessageHandler.UploadMetadataProvider() {
-
-				@Override
-				public void populateMetadata(ObjectMetadata metadata, Message<?> message) {
-					if (message.getPayload() instanceof InputStream) {
-						metadata.setContentLength(1);
-						metadata.setContentType(MediaType.APPLICATION_JSON_VALUE);
-						metadata.setContentDisposition("test.json");
-					}
+			s3MessageHandler.setUploadMetadataProvider((metadata, message) -> {
+				if (message.getPayload() instanceof InputStream) {
+					metadata.setContentLength(1);
+					metadata.setContentType(MediaType.APPLICATION_JSON_VALUE);
+					metadata.setContentDisposition("test.json");
 				}
-
 			});
 			s3MessageHandler.setProgressListener(s3ProgressListener());
 			return s3MessageHandler;
