@@ -1,17 +1,17 @@
 /*
- * Copyright 2015-2016 the original author or authors.
+ * Copyright 2016 the original author or authors.
  *
- *  Licensed under the Apache License, Version 2.0 (the "License");
- *  you may not use this file except in compliance with the License.
- *  You may obtain a copy of the License at
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
  *
- *       http://www.apache.org/licenses/LICENSE-2.0
+ *      http://www.apache.org/licenses/LICENSE-2.0
  *
- *  Unless required by applicable law or agreed to in writing, software
- *  distributed under the License is distributed on an "AS IS" BASIS,
- *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  See the License for the specific language governing permissions and
- *  limitations under the License.
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
  */
 
 package org.springframework.integration.aws.inbound;
@@ -199,11 +199,11 @@ public class SnsInboundChannelAdapter extends HttpRequestHandlingMessagingGatewa
 
 	private static class NotificationStatusResolver extends NotificationStatusHandlerMethodArgumentResolver {
 
-		public NotificationStatusResolver(AmazonSNS amazonSns) {
+		NotificationStatusResolver(AmazonSNS amazonSns) {
 			super(amazonSns);
 		}
 
-		protected NotificationStatus resolveNotificationStatus(JsonNode content) {
+		NotificationStatus resolveNotificationStatus(JsonNode content) {
 			return (NotificationStatus) doResolveArgumentFromNotificationMessage(content, null, null);
 		}
 
