@@ -1,17 +1,17 @@
 /*
- * Copyright 2015 the original author or authors.
+ * Copyright 2016 the original author or authors.
  *
- *  Licensed under the Apache License, Version 2.0 (the "License");
- *  you may not use this file except in compliance with the License.
- *  You may obtain a copy of the License at
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
  *
- *       http://www.apache.org/licenses/LICENSE-2.0
+ *      http://www.apache.org/licenses/LICENSE-2.0
  *
- *  Unless required by applicable law or agreed to in writing, software
- *  distributed under the License is distributed on an "AS IS" BASIS,
- *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  See the License for the specific language governing permissions and
- *  limitations under the License.
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
  */
 
 package org.springframework.integration.aws.support;
@@ -29,7 +29,7 @@ import org.springframework.util.Assert;
  *
  * @author Artem Bilan
  */
-public class SnsBodyBuilder {
+public final class SnsBodyBuilder {
 
 	private final Map<String, String> snsMessage = new HashMap<>();
 
@@ -50,7 +50,7 @@ public class SnsBodyBuilder {
 
 	public String build() {
 		StringBuilder stringBuilder = new StringBuilder("{");
-		for (Map.Entry<String, String> entry : snsMessage.entrySet()) {
+		for (Map.Entry<String, String> entry : this.snsMessage.entrySet()) {
 			stringBuilder.append("\"")
 					.append(entry.getKey())
 					.append("\":\"")
