@@ -16,11 +16,7 @@
 
 package org.springframework.integration.aws.config.xml;
 
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertFalse;
-import static org.junit.Assert.assertNotNull;
-import static org.junit.Assert.assertSame;
-import static org.junit.Assert.assertTrue;
+import static org.assertj.core.api.Assertions.assertThat;
 
 import java.lang.reflect.Method;
 import java.util.Collection;
@@ -70,35 +66,37 @@ public class S3InboundChannelAdapterParserTests {
 
 	@Test
 	public void testFtpInboundChannelAdapterComplete() throws Exception {
-		assertFalse(TestUtils.getPropertyValue(this.s3Inbound, "autoStartup", Boolean.class));
+		assertThat(TestUtils.getPropertyValue(this.s3Inbound, "autoStartup", Boolean.class)).isFalse();
 		PriorityBlockingQueue<?> blockingQueue = TestUtils.getPropertyValue(this.s3Inbound,
 				"source.fileSource.toBeReceived", PriorityBlockingQueue.class);
 		Comparator<?> comparator = blockingQueue.comparator();
-		assertSame(this.comparator, comparator);
-		assertEquals("s3Inbound", this.s3Inbound.getComponentName());
-		assertEquals("aws:s3-inbound-channel-adapter", this.s3Inbound.getComponentType());
-		assertSame(this.s3Channel, TestUtils.getPropertyValue(this.s3Inbound, "outputChannel"));
+		assertThat(comparator).isSameAs(this.comparator);
+		assertThat(this.s3Inbound.getComponentName()).isEqualTo("s3Inbound");
+		assertThat(this.s3Inbound.getComponentType()).isEqualTo("aws:s3-inbound-channel-adapter");
+		assertThat(TestUtils.getPropertyValue(this.s3Inbound, "outputChannel")).isSameAs(this.s3Channel);
 
 		S3InboundFileSynchronizingMessageSource inbound = TestUtils.getPropertyValue(this.s3Inbound, "source",
 				S3InboundFileSynchronizingMessageSource.class);
 
 		S3InboundFileSynchronizer fisync = TestUtils.getPropertyValue(inbound, "synchronizer",
 				S3InboundFileSynchronizer.class);
-		assertEquals("'foo/bar'",
-				TestUtils.getPropertyValue(fisync, "remoteDirectoryExpression", Expression.class)
-						.getExpressionString());
-		assertNotNull(TestUtils.getPropertyValue(fisync, "localFilenameGeneratorExpression"));
-		assertTrue(TestUtils.getPropertyValue(fisync, "preserveTimestamp", Boolean.class));
-		assertEquals(".foo", TestUtils.getPropertyValue(fisync, "temporaryFileSuffix", String.class));
+		assertThat(TestUtils.getPropertyValue(fisync, "remoteDirectoryExpression", Expression.class)
+				.getExpressionString())
+				.isEqualTo("'foo/bar'");
+		assertThat(TestUtils.getPropertyValue(fisync, "localFilenameGeneratorExpression")).isNotNull();
+		assertThat(TestUtils.getPropertyValue(fisync, "preserveTimestamp", Boolean.class)).isTrue();
+		assertThat(TestUtils.getPropertyValue(fisync, "temporaryFileSuffix", String.class)).isEqualTo(".foo");
 		String remoteFileSeparator = (String) TestUtils.getPropertyValue(fisync, "remoteFileSeparator");
-		assertNotNull(remoteFileSeparator);
-		assertEquals("", remoteFileSeparator);
+		assertThat(remoteFileSeparator).isNotNull();
+		assertThat(remoteFileSeparator).isEmpty();
 		S3SimplePatternFileListFilter filter = TestUtils.getPropertyValue(fisync, "filter",
 				S3SimplePatternFileListFilter.class);
-		assertNotNull(filter);
-		assertSame(this.s3SessionFactory, TestUtils.getPropertyValue(fisync, "remoteFileTemplate.sessionFactory"));
-		assertTrue(TestUtils.getPropertyValue(inbound, "fileSource.scanner.filter.fileFilters", Collection.class)
-				.contains(this.acceptAllFilter));
+		assertThat(filter).isNotNull();
+		assertThat(TestUtils.getPropertyValue(fisync, "remoteFileTemplate.sessionFactory"))
+				.isSameAs(this.s3SessionFactory);
+		assertThat(TestUtils.getPropertyValue(inbound, "fileSource.scanner.filter.fileFilters", Collection.class)
+				.contains(this.acceptAllFilter))
+				.isTrue();
 		final AtomicReference<Method> genMethod = new AtomicReference<Method>();
 		ReflectionUtils.doWithMethods(AbstractInboundFileSynchronizer.class, new ReflectionUtils.MethodCallback() {
 
@@ -111,7 +109,7 @@ public class S3InboundChannelAdapterParserTests {
 			}
 
 		});
-		assertEquals("FOO.afoo", genMethod.get().invoke(fisync, "foo"));
+		assertThat(genMethod.get().invoke(fisync, "foo")).isEqualTo("FOO.afoo");
 	}
 
 }
