@@ -1,29 +1,26 @@
 /*
- * Copyright 2015 the original author or authors.
+ * Copyright 2015-2016 the original author or authors.
  *
- *  Licensed under the Apache License, Version 2.0 (the "License");
- *  you may not use this file except in compliance with the License.
- *  You may obtain a copy of the License at
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
  *
- *       http://www.apache.org/licenses/LICENSE-2.0
+ *      http://www.apache.org/licenses/LICENSE-2.0
  *
- *  Unless required by applicable law or agreed to in writing, software
- *  distributed under the License is distributed on an "AS IS" BASIS,
- *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  See the License for the specific language governing permissions and
- *  limitations under the License.
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
  */
 
 package org.springframework.integration.aws.inbound;
 
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertNotNull;
-import static org.junit.Assert.assertThat;
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.mockito.BDDMockito.given;
+import static org.mockito.BDDMockito.mock;
 import static org.mockito.Matchers.any;
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.when;
 
-import org.hamcrest.Matchers;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
@@ -60,13 +57,13 @@ public class SqsMessageDrivenChannelAdapterTests {
 	@Test
 	public void testSqsMessageDrivenChannelAdapter() {
 		org.springframework.messaging.Message<?> receive = this.inputChannel.receive(1000);
-		assertNotNull(receive);
-		assertThat((String) receive.getPayload(), Matchers.isOneOf("messageContent", "messageContent2"));
-		assertEquals("testQueue", receive.getHeaders().get(AwsHeaders.QUEUE));
+		assertThat(receive).isNotNull();
+		assertThat((String) receive.getPayload()).isIn("messageContent", "messageContent2");
+		assertThat(receive.getHeaders().get(AwsHeaders.QUEUE)).isEqualTo("testQueue");
 		receive = this.inputChannel.receive(1000);
-		assertNotNull(receive);
-		assertThat((String) receive.getPayload(), Matchers.isOneOf("messageContent", "messageContent2"));
-		assertEquals("testQueue", receive.getHeaders().get(AwsHeaders.QUEUE));
+		assertThat(receive).isNotNull();
+		assertThat((String) receive.getPayload()).isIn("messageContent", "messageContent2");
+		assertThat(receive.getHeaders().get(AwsHeaders.QUEUE)).isEqualTo("testQueue");
 	}
 
 	@Configuration
@@ -76,20 +73,22 @@ public class SqsMessageDrivenChannelAdapterTests {
 		@Bean
 		public AmazonSQSAsync amazonSqs() {
 			AmazonSQSAsync sqs = mock(AmazonSQSAsync.class);
-			when(sqs.getQueueUrl(new GetQueueUrlRequest("testQueue"))).thenReturn(new GetQueueUrlResult().
-					withQueueUrl("http://testQueue.amazonaws.com"));
+			given(sqs.getQueueUrl(new GetQueueUrlRequest("testQueue")))
+					.willReturn(new GetQueueUrlResult().withQueueUrl("http://testQueue.amazonaws.com"));
 
 
-			when(sqs.receiveMessage(new ReceiveMessageRequest("http://testQueue.amazonaws.com")
+			given(sqs.receiveMessage(new ReceiveMessageRequest("http://testQueue.amazonaws.com")
 					.withAttributeNames("All")
 					.withMessageAttributeNames("All")
 					.withMaxNumberOfMessages(10)))
-					.thenReturn(new ReceiveMessageResult()
+					.willReturn(new ReceiveMessageResult()
 							.withMessages(new Message().withBody("messageContent"),
 									new Message().withBody("messageContent2")))
-					.thenReturn(new ReceiveMessageResult());
-			when(sqs.getQueueAttributes(any(GetQueueAttributesRequest.class)))
-					.thenReturn(new GetQueueAttributesResult());
+					.willReturn(new ReceiveMessageResult());
+
+			given(sqs.getQueueAttributes(any(GetQueueAttributesRequest.class)))
+					.willReturn(new GetQueueAttributesResult());
+
 			return sqs;
 		}
 
